adjustments after CR
