+++ widgets/blueprintCatalog/src/RepositoryLinkButton.tsx
@@ -1,0 +1,47 @@
+import styled from 'styled-components';
+import Utils from './utils';
+
+interface IconLinkProps {
+    url: string;
+}
+const { Button } = Stage.Basic;
+
+const StyledLinkButton = styled(Button)`
+    &&&& {
+        padding: 0;
+        margin-right: 10px;
+        font-size: 22px;
+        background: none;
+        &:hover {
+            color: #1b1f23;
+        }
+    }
+`;
+
+const t = Utils.getWidgetTranslation('blueprintCatalog');
+
+const getIcon = (url: string): string => {
+    const isGithub = url.startsWith('https://github.com');
+    const isGitlab = url.startsWith('https://gitlab.com');
+    const isBitbucket = url.startsWith('https://bitbucket.org');
+
+    if (isGithub) return 'github';
+    if (isBitbucket) return 'bitbucket';
+    if (isGitlab) return 'gitlab';
+    return 'git';
+};
+
+const RepositoryLinkButton = ({ url }: IconLinkProps) => {
+    const icon = getIcon(url);
+
+    return (
+        <StyledLinkButton
+            circular
+            icon={icon}
+            onClick={() => Stage.Utils.Url.redirectToPage(url)}
+            title={t('actions.openBlueprintRepository')}
+        />
+    );
+};
+
+export default RepositoryLinkButton;
